requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1

# LlamaIndex & AI
llama-index==0.12.11
llama-index-llms-groq==0.3.1
llama-index-embeddings-huggingface==0.5.1

# Embeddings & Transformers
sentence-transformers==3.3.1
torch==2.5.1

# Security & Validation
python-multipart==0.0.20

# Git Repository Handling
GitPython==3.1.43
